refactor(api): deduplicate AI action setup
