+++ packages/api/src/extensions/actions/ai/generate-object.base.action.ts
@@ -4,7 +4,7 @@
  * Full terms: see LICENSE.md.
  */
 
-import { JSONSchema7, Output, ToolSet, generateText, jsonSchema } from 'ai';
+import { JSONSchema7, Output, generateText, jsonSchema } from 'ai';
 
 import { ActionService } from '@/actions/actions.service';
 import { ActionMetadata, ExecArgs } from '@/actions/types';
@@ -30,55 +30,20 @@ export abstract class AiGenerateObjectBaseAction<
 
   protected abstract resolvePromptInput(input: I): AiPromptInput;
 
-  async execute({
-    input,
-    settings,
-    context,
-    bindings,
-    signal,
-  }: ExecArgs<I, C, AiGenerateObjectSettings>) {
-    const logger = context.services.logger;
-    const modelBinding = bindings.model;
-    const providerName = modelBinding?.settings?.provider ?? 'openai';
-    const modelId = this.resolveModelId(modelBinding);
-    const credentials = await context.services.credentials.findOneValue(
-      modelBinding?.settings?.api_key,
-    );
-    const providerOptions = this.buildProviderInitOptions(
-      providerName,
-      modelBinding,
-      credentials,
-    );
-    const provider = await this.loadProvider(providerName, providerOptions);
-    const model = this.createModel(provider, modelId);
-    const selectedMemorySlugs = this.resolveMemoryBindingSlugs(
-      context,
-      bindings.memory,
-    );
-    const promptPayload = await this.buildPrompt(
-      this.resolvePromptInput(input),
-      context,
-      selectedMemorySlugs,
-    );
-    const callSettings = this.buildCallSettings(settings);
+  async execute(args: ExecArgs<I, C, AiGenerateObjectSettings>) {
+    const { input, settings, signal } = args;
+    const {
+      callSettings,
+      logCall,
+      model,
+      modelId,
+      promptPayload,
+      stopWhen,
+      tools,
+    } = await this.prepareCall(this.resolvePromptInput(input), args);
     // Structured outputs do not support stop sequences in the AI SDK call.
     const { stopSequences: _stopSequences, ...callSettingsWithoutStops } =
       callSettings;
-    const tools = (await this.buildTools(
-      context,
-      bindings.tools,
-      bindings.mcp,
-      selectedMemorySlugs,
-      signal,
-    )) as ToolSet | undefined;
-    const toolNames = [
-      ...Object.keys(bindings.tools ?? {}),
-      ...Object.keys(bindings.mcp ?? {}),
-    ];
-    const { stopWhen, stepCount, toolCall } = this.buildStopWhen(
-      settings,
-      tools,
-    );
     const outputSchema = settings.output_schema as JSONSchema7;
     const output = Output.object({
       schema: jsonSchema(outputSchema),
@@ -86,18 +51,7 @@ export abstract class AiGenerateObjectBaseAction<
       description: outputSchema.description,
     });
 
-    logger.debug(
-      `Calling model "${modelId}" via ${this.name} action using provider "${providerName}"`,
-      {
-        provider: providerName,
-        base_url: providerOptions.baseURL,
-        tools: toolNames,
-        stop_when: {
-          step_count: stepCount,
-          tool_call: toolCall,
-        },
-      },
-    );
+    logCall();
     const result = await generateText({
       ...promptPayload,
       ...callSettingsWithoutStops,
